Resolved a lot of compiler warnings (-Wall) and added more code for polygon stuff (still not verified, though. Also, I should probably check the output of physics vectors to the screen as it seems to be incorrect

+++ src/System.cpp
@@ -55,7 +55,7 @@ int System::readfile(SystemClient* client, const char* path, size_t start, size_
 
         }
         //Get to the starting position
-        if (f.seek(long(start), SEEK_SET) != start)
+        if (f.seek(long(start), SEEK_SET) != (long int)start)
         {
             //Error
         }
